+++ src/main/java/com/cashonline/loanmanagementsystem/persistence/dao/LoanRepository.java
@@ -3,5 +3,5 @@ package com.cashonline.loanmanagementsystem.persistence.dao;
 import com.cashonline.loanmanagementsystem.persistence.entities.LoanEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LoanRepository extends JpaRepository<LoanEntity, Long> {
+public interface LoanRepository extends JpaRepository<LoanEntity, Integer> {
 }
