+++ src/main/java/com/cashonline/loanmanagementsystem/persistence/dao/PersonRepository.java
@@ -3,6 +3,6 @@ package com.cashonline.loanmanagementsystem.persistence.dao;
 import com.cashonline.loanmanagementsystem.persistence.entities.PersonEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonRepository extends JpaRepository<PersonEntity, Long> {
+public interface PersonRepository extends JpaRepository<PersonEntity, Integer> {
 
 }
